### 2025-04-16 #8059 @editorialbot
[PRE REVIEW]: Ecos: An accessible and intuitive co-simulation framework
**Submitting author:** @markaren (<a href="http://orcid.org/0000-0001-6436-7213">Lars Ivar Hatledal</a>)
**Repository:** https://github.com/Ecos-platform/ecos
**Branch with paper.md** (empty if default branch): joss
**Version:** 0.5.5
