### 2025-04-27 64e4fbb
bool must be checked before int #46 [skip ci]
