### 2025-04-27 8f1d884
enable fmiLogging from resolvers and pass to ecos logger
replace printf
