### 2025-04-27 c72634f
pass fmiLogging to ecos logger (#47)
