### 2025-04-29 #8059 @espottesmith
Hey @crvernon. Yep, I can take this one on.

@markaren, it'd be helpful if you provided a list of potential reviewers. You can look at the authors of the suggested papers listed above or use our "[search for reviewers](https://reviewers.joss.theoj.org/reviewers)" tool.

I took at look at your paper, and while you do seem to meet all of our criteria, the descriptions are rather bare-bones at the moment. In particular, I'd love to see a more detailed comparison between ecos and other similar tools.
