### 2025-04-29 #8059 @markaren
Possible reviewers:  @elac-safran, @lgiannantoni, @parikshitbajpai, @ElektrikAkar

> I took at look at your paper, and while you do seem to meet all of our criteria, the descriptions are rather bare-bones at the moment. In particular, I'd love to see a more detailed comparison between ecos and other similar tools.

I will look into it.
