### 2025-04-29 94052d5
fix error in README example [skip ci]
