### 2025-04-29 6eaf9bf
make it possible to configure cpp plot programmatically
