### 2025-04-29 0af4ab7
Make it possible to configure cpp plot programmatically (#49)
