+++ src/ecos/util/plotter.cpp
@@ -1,6 +1,8 @@
 
 #include "ecos/util/plotter.hpp"
 
+#include "util/temp_dir.hpp"
+
 #include "ecos/logger/logger.hpp"
 
 #include <fstream>
@@ -10,7 +12,159 @@
 
 using namespace ecos;
 
+namespace
+{
+
+// --- XML Serialization Helpers ---
+
+std::string escape(const std::string& str)
+{
+    std::string out;
+    for (char c : str) {
+        switch (c) {
+            case '&': out += "&amp;"; break;
+            case '<': out += "&lt;"; break;
+            case '>': out += "&gt;"; break;
+            case '\"': out += "&quot;"; break;
+            case '\'': out += "&apos;"; break;
+            default: out += c;
+        }
+    }
+    return out;
+}
+
+void indent(std::ostream& os, int level)
+{
+    os << std::string(level * 2, ' ');
+}
+
+void writeVariableIdentifier(std::ostream& os, const std::string& tag, const variable_identifier& vi, int indentLevel)
+{
+    indent(os, indentLevel);
+    os << "<" << tag
+       << " component=\"" << escape(vi.instanceName) << "\""
+       << " variable=\"" << escape(vi.variableName) << "\"/>\n";
+}
+
+void writeXYSeries(std::ostream& os, const TXYSeries& series, int indentLevel)
+{
+    indent(os, indentLevel);
+    os << "<ecos:series name=\"" << escape(series.name) << "\"";
+    if (series.marker)
+        os << " marker=\"" << escape(*series.marker) << "\"";
+    os << ">\n";
+
+    writeVariableIdentifier(os, "x", series.x, indentLevel + 1);
+    writeVariableIdentifier(os, "y", series.y, indentLevel + 1);
+
+    indent(os, indentLevel);
+    os << "</ecos:series>\n";
+}
+
+void writeXySeriesChart(std::ostream& os, const TXYSeriesChart& chart, int indentLevel)
+{
+    indent(os, indentLevel);
+    os << "<ecos:xyseries title=\"" << escape(chart.title)
+       << "\" xLabel=\"" << escape(chart.xLabel)
+       << "\" yLabel=\"" << escape(chart.yLabel) << "\">\n";
+
+    for (const auto& series : chart.series) {
+        writeXYSeries(os, series, indentLevel + 1);
+    }
+
+    indent(os, indentLevel);
+    os << "</ecos:xyseries>\n";
+}
+
+void writeLinearTransformation(std::ostream& os, const TLinearTransformation& lt, int indentLevel)
+{
+    indent(os, indentLevel);
+    os << "<ecos:linearTransformation offset=\"" << lt.offset << "\" factor=\"" << lt.factor << "\"/>\n";
+}
+
+void writeVariable(std::ostream& os, const TVariable& var, int indentLevel)
+{
+    indent(os, indentLevel);
+    os << "<ecos:variable name=\"" << escape(var.name) << "\">\n";
+    if (var.linearTransformation) {
+        writeLinearTransformation(os, *var.linearTransformation, indentLevel + 1);
+    }
+    indent(os, indentLevel);
+    os << "</ecos:variable>\n";
+}
+
+void writeComponent(std::ostream& os, const TComponent& comp, int indentLevel)
+{
+    indent(os, indentLevel);
+    os << "<ecos:component name=\"" << escape(comp.name) << "\">\n";
+    for (const auto& var : comp.variables) {
+        writeVariable(os, var, indentLevel + 1);
+    }
+    indent(os, indentLevel);
+    os << "</ecos:component>\n";
+}
+
[…]
