+++ examples/CMakeLists.txt
@@ -17,6 +17,7 @@ endif ()
 
 
 add_example_executable(dp_ship dp-ship/dp_ship.cpp)
+add_example_executable(bouncing_ball bouncing_ball/bouncing_ball.cpp)
 
 if (WIN32)
     add_example_executable(quarter_truck quarter_truck/quarter_truck.cpp)
