### 2025-05-01 2bf3d31
add basic tests
