### 2025-05-01 7c2ba5f
run pytests on CI
