### 2025-05-01 a45af7a
install ecospy
