+++ .github/workflows/build.yml
@@ -82,6 +82,7 @@ jobs:
       - name: Install Python dependencies
         run: |
           python -m pip install --upgrade pip
+          pip install .
           pip install pytest
 
       - name: Run Python tests
