+++ .github/workflows/build.yml
@@ -67,6 +67,7 @@ jobs:
           path: |
             build/bin/ecos*
             build/bin/proxyfmu*
+            build/bin/libecosc.*
 
   python-tests:
     runs-on: ubuntu-22.04
@@ -78,6 +79,12 @@ jobs:
       - uses: actions/setup-python@v5
         with:
           python-version: '3.11'
+
+      - name: Download ecospy binaries
+        uses: actions/download-artifact@v4
+        with:
+          name: ecos-${{ matrix.os }}
+          path: ecospy/binaries
 
       - name: Install Python dependencies
         run: |
