### 2025-05-01 103d7b2
fix name and upload shared library on linux
