+++ .github/workflows/build.yml
@@ -36,6 +36,7 @@ jobs:
           path: |
             build/bin/ecos*
             build/bin/proxyfmu*
+            build/bin/libecosc.*
 
   cmake-on-windows:
 
@@ -83,7 +84,7 @@ jobs:
       - name: Download ecospy binaries
         uses: actions/download-artifact@v4
         with:
-          name: ecos-${{ matrix.os }}
+          name: ecos-ubuntu-22.04
           path: ecospy/binaries
 
       - name: Install Python dependencies
