+++ .github/workflows/build.yml
@@ -77,6 +77,10 @@ jobs:
 
     steps:
       - uses: actions/checkout@v4
+      - name: Install prerequisites
+        run: |
+          sudo apt install libtbb-dev
+
       - uses: actions/setup-python@v5
         with:
           python-version: '3.11'
