+++ .github/workflows/build.yml
@@ -36,6 +36,7 @@ jobs:
           path: |
             build/bin/ecos*
             build/bin/proxyfmu*
+            build/bin/libecosc.*
 
   cmake-on-windows:
 
@@ -67,3 +68,35 @@ jobs:
           path: |
             build/bin/ecos*
             build/bin/proxyfmu*
+            build/bin/libecosc.*
+
+  python-tests:
+    runs-on: ubuntu-22.04
+    timeout-minutes: 10
+    needs: cmake-on-linux  # only run after cmake-on-linux succeeds
+
+    steps:
+      - uses: actions/checkout@v4
+      - name: Install prerequisites
+        run: |
+          sudo apt install libtbb-dev
+
+      - uses: actions/setup-python@v5
+        with:
+          python-version: '3.11'
+
+      - name: Download ecospy binaries
+        uses: actions/download-artifact@v4
+        with:
+          name: ecos-ubuntu-22.04
+          path: ecospy/binaries
+
+      - name: Install Python dependencies
+        run: |
+          python -m pip install --upgrade pip
+          pip install .
+          pip install pytest
+
+      - name: Run Python tests
+        run: |
+          pytest ecospy/tests -v
