### 2025-05-04 d11e98c
expose simulation_runner
