### 2025-05-04 3ce751b
expose simulation_runner (#51)
