### 2025-05-05 2324ee1
add van_der_pol cpp example [skip ci]
