+++ doc/paper/paper.bib
@@ -13,7 +13,8 @@
   booktitle={9th International Modelica Conference},
   pages={173--184},
   year={2012},
-  organization={The Modelica Association}
+  organization={The Modelica Association},
+  doi={10.3384/ecp12076173},
 }
 
 @inproceedings{junghanns2021functional,
@@ -21,7 +22,8 @@
   author={Junghanns, Andreas and Gomes, Cl{\'a}udio and Schulze, Christian and Schuch, Klaus and Pierre, R and Blaesken, Matthias and Zacharias, Irina and Pillekeit, Andreas and Wernersson, Karl and Sommer, Torsten and others},
   booktitle={Modelica conferences},
   pages={17--26},
-  year={2021}
+  year={2021},
+  doi={10.3384/ecp2118117},
 }
 
 @misc{FMPy,
@@ -46,7 +48,8 @@
   volume={108},
   pages={102243},
   year={2021},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.simpat.2021.102243}
 }
 
 @inproceedings{ochel2019omsimulator,
