+++ doc/paper/paper.bib
@@ -1,11 +1,3 @@
-@inproceedings{blochwitz2011functional,
-  title={The functional mockup interface for tool independent exchange of simulation models},
-  author={Blochwitz, Torsten and Otter, Martin and Arnold, Martin and Bausch, Constanze and Clau{\ss}, Christoph and Elmqvist, Hilding and Junghanns, Andreas and Mauss, Jakob and Monteiro, Manuel and Neidhold, Thomas and others},
-  booktitle={Proceedings of the 8th international Modelica conference},
-  pages={105--114},
-  year={2011},
-  organization={Link{\"o}ping University Press}
-}
 
 @inproceedings{blochwitz2012functional,
   title={Functional mockup interface 2.0: The standard for tool independent exchange of simulation models},
@@ -38,7 +30,8 @@
 @article{kohler2016modelica,
   title={Modelica-association-project “system structure and parameterization”--early insights},
   author={K{\"o}hler, Jochen and Heinkel, Hans-Martin and Mai, Pierre and Krasser, J{\"u}rgen and Deppe, Markus and Nagasawa, Mikio},
-  year={2016}
+  year={2016},
+  doi={10.3384/ecp1612435}
 }
 
 @article{hatledal2021vico,
@@ -57,7 +50,8 @@
   author={Ochel, Lennart A and Braun, Robert and Thiele, Bernhard and Asghar, Adeel and Buffoni, Lena and Eek, Magnus and Fritzson, Peter and Fritzson, Dag and Horkeby, Sune and H{\"a}llquist, Robert and others},
   booktitle={Modelica},
   pages={157--007},
-  year={2019}
+  year={2019},
+  doi={10.3384/ecp1915769}
 }
 
 @inproceedings{smogeli2020open,
