### 2025-05-07 657c295
add proxy example for bouncing_ball
