### 2025-05-07 10f57b6
make it possible to set csv decimationfactor from C/Python
