+++ include/ecos/ecos.h
@@ -95,6 +95,7 @@ ecos_simulation_listener_t* ecos_simulation_listener_create(ecos_simulation_list
 void ecos_simulation_add_listener(ecos_simulation_t* sim, const char* name, ecos_simulation_listener_t* listener);
 void ecos_simulation_remove_listener(ecos_simulation_t* sim, const char* name);
 ecos_simulation_listener_t* ecos_csv_writer_create(const char* resultFile, const char* csvConfig = nullptr);
+bool ecos_csv_writer_set_decimation_factor(ecos_simulation_listener_t* writer, int decimationFactor);
 
 void ecos_plot_csv(const char* csvFile, const char* chartConfig);
 // -------------
