+++ ecospy/EcosSimulation.py
@@ -142,12 +142,20 @@ class EcosSimulation:
     def remove_listener(self, name: str):
         self._remove_listener(self.sim, name.encode())
 
-    def add_csv_writer(self, result_file: str, csv_config: str = None):
+    def add_csv_writer(self, result_file: str, csv_config: str = None, decimationFactor: int = None):
 
         listener = self._create_csv_writer(result_file.encode(),
                                            None if csv_config is None else csv_config.encode())
         if listener is None:
             raise Exception(EcosLib.get_last_error())
+
+        if decimationFactor is not None:
+            set_decimation = dll.ecos_csv_writer_set_decimation_factor
+            set_decimation.argtypes = [c_void_p, c_int]
+            set_decimation.restype = c_bool
+
+            if not set_decimation(listener, decimationFactor):
+                raise Exception(EcosLib.get_last_error())
 
         self._add_listener(self.sim, b'csv_writer', listener)
 
