### 2025-05-08 85abdaa
expand python tests with proxyfmu
