### 2025-05-08 23fd214
improve proxymfu error handling
