### 2025-05-09 f1d2345
fix segfault in fmi3 [skip ci]
