+++ src/proxyfmu/proxy_slave.cpp
@@ -21,7 +21,7 @@ using namespace simple_socket;
 namespace
 {
 
-std::string read_data(std::string const& fileName)
+std::vector<uint8_t> read_data(std::string const& fileName)
 {
     std::ifstream file(fileName, std::ios::binary | std::ios::ate);
     if (!file) {
@@ -30,12 +30,12 @@ std::string read_data(std::string const& fileName)
     const std::streamsize size = file.tellg();
     file.seekg(0, std::ios::beg);
 
-    std::vector<char> buffer(size);
-    if (!file.read(buffer.data(), size)) {
+    std::vector<uint8_t> buffer(size);
+    if (!file.read(reinterpret_cast<char*>(buffer.data()), size)) {
         throw std::runtime_error("Failed to read file: " + fileName);
     }
 
-    return {buffer.begin(), buffer.end()};
+    return buffer;
 }
 
 } // namespace
@@ -74,13 +74,15 @@ proxy_slave::proxy_slave(
             throw std::runtime_error("Failed to connect to: " + address);
         }
 
-        const std::string data = read_data(fmuPath.string());
+        const auto data = read_data(fmuPath.string());
         const std::string fmuName = std::filesystem::path(fmuPath).stem().string();
 
         flexbuffers::Builder fbb;
-        fbb.String(fmuName);
-        fbb.String(instanceName);
-        fbb.String(data);
+        fbb.Vector([&]() {
+            fbb.String(fmuName);
+            fbb.String(instanceName);
+            fbb.Blob(data); // no key here in a vector
+        });
         fbb.Finish();
 
         const auto msgLen = fbb.GetSize();
