### 2025-05-09 66a5608
add binaries to PATH
