+++ ecospy/lib.py
@@ -2,20 +2,29 @@ import os
 from pathlib import Path
 from ctypes import CDLL, Structure, c_char_p, c_int
 
+
 def load_library():
     def suffix() -> str:
         return ".dll" if os.name == "nt" else ".so"
 
-    bin_folder = (Path(__file__).parent / 'binaries').resolve()
+    bin_folder = str((Path(__file__).parent / 'binaries').resolve())
+    lib_name = f"libecosc{suffix()}"
+
+    current_path = os.environ.get("PATH", "")
+
+    # Add the executable directory to the PATH if not already present
+    if bin_folder not in current_path:
+        os.environ["PATH"] = bin_folder + os.pathsep + current_path
 
     if os.name == "nt":
-        with os.add_dll_directory(str(bin_folder)):
-            return CDLL(f"libecosc{suffix()}")
+        with os.add_dll_directory(bin_folder):
+            return CDLL(lib_name)
     else:
-        return CDLL(f"{str(bin_folder)}/libecosc{suffix()}")
+        return CDLL(f"{bin_folder}/libecosc{suffix()}")
 
 
 dll = load_library()
+
 
 class Version(Structure):
     _fields_ = [("major", c_int), ("minor", c_int), ("patch", c_int)]
