+++ doc/paper/paper.bib
@@ -42,7 +42,7 @@
   pages={102243},
   year={2021},
   publisher={Elsevier},
-  doi={10.1016/j.simpat.2021.102243}
+  doi={10.1016/j.simpat.2020.102243}
 }
 
 @inproceedings{ochel2019omsimulator,
