### 2025-05-09 da6e4ea
Fallback to no_op plotter if matplotlib is not available (#55)
