+++ ecospy/plotter.py
@@ -1,8 +1,16 @@
-import re
-
-import pandas as pd
-import matplotlib.pyplot as plt
 from typing import Callable
+
+try:
+    import matplotlib.pyplot as plt
+
+    MATPLOTLIB_AVAILABLE = True
+
+    import re
+    import pandas as pd
+
+
+except ImportError:
+    MATPLOTLIB_AVAILABLE = False
 
 
 class TimeSeriesConfig:
@@ -24,76 +32,94 @@ class XYSeriesConfig:
         self.series = series
 
 
-class Plotter:
-    fig_id = 0
+if MATPLOTLIB_AVAILABLE:
+    class Plotter:
+        fig_id = 0
 
-    def __init__(self, csv_file, configs):
-        self.csv = pd.read_csv(csv_file, delimiter=r",\s+", engine="python")
+        def __init__(self, csv_file, configs):
+            self.csv = pd.read_csv(csv_file, delimiter=r",\s+", engine="python")
 
-        if not isinstance(configs, (list, tuple)):
-            configs = [configs]
+            if not isinstance(configs, (list, tuple)):
+                configs = [configs]
 
-        for config in configs:
-            if isinstance(config, TimeSeriesConfig):
-                self.make_time_series(self.csv, config)
-            elif isinstance(config, XYSeriesConfig):
-                self.make_xy_series(self.csv, config)
-            else:
-                raise Exception("Missing or unsupported configuration")
+            for config in configs:
+                if isinstance(config, TimeSeriesConfig):
+                    self.make_time_series(self.csv, config)
+                elif isinstance(config, XYSeriesConfig):
+                    self.make_xy_series(self.csv, config)
+                else:
+                    raise Exception("Missing or unsupported configuration")
 
-    @staticmethod
-    def show():
-        plt.show()
+        @staticmethod
+        def show():
+            plt.show()
 
-    @staticmethod
-    def make_time_series(csv, timeseries: TimeSeriesConfig):
+        @staticmethod
+        def make_time_series(csv, timeseries: TimeSeriesConfig):
 
-        t = csv['time']
-        plt.figure("figure_{}".format(Plotter.fig_id))
-        Plotter.fig_id += 1
-        plt.title(timeseries.title)
-        plt.xlabel("time[s]")
-        plt.ylabel(timeseries.y_label)
+            t = csv['time']
+            plt.figure("figure_{}".format(Plotter.fig_id))
+            Plotter.fig_id += 1
+            plt.title(timeseries.title)
+            plt.xlabel("time[s]")
+            plt.ylabel(timeseries.y_label)
 
-        # Create a mapping from cleaned column names to actual column names
-        clean_col_map = {}
-        for col in csv.columns:
-            clean_col = re.sub(r'\s*\[.*?]\s*$', '', col)  # Remove trailing type markers like [REAL], [INT]
-            clean_col_map[clean_col] = col
+            # Create a mapping from cleaned column names to actual column names
+            clean_col_map = {}
+            for col in csv.columns:
+                clean_col = re.sub(r'\s*\[.*?]\s*$', '', col)  # Remove trailing type markers like [REAL], [INT]
+                clean_col_map[clean_col] = col
 
-        for identifier in timeseries.identifiers:
-            if identifier in clean_col_map:
-                actual_col = clean_col_map[identifier]
-                data = csv[actual_col]
-                if timeseries.modifiers is not None and identifier in timeseries.modifiers:
-                    data = data.map(timeseries.modifiers[identifier])
-                plt.plot(t, data, label=actual_col)
-            else:
-                print(f"Warning: Identifier '{identifier}' not found in CSV columns.")
+            for identifier in timeseries.identifiers:
[…]
