### 2025-05-11 a8edf97
fix proxyfmu issue
when running under python under nix
