### 2025-05-11 cf5ad14
make proxyfmu executable in CI
