### 2025-05-12 e666314
Expand python tests with proxyfmu (#53)
