+++ examples/bouncing_ball/bouncing_ball_proxy.cpp
@@ -6,6 +6,8 @@
 #include <ecos/logger/logger.hpp>
 #include <ecos/util/plotter.hpp>
 
+#include <iostream>
+
 using namespace ecos;
 
 int main()
@@ -16,21 +18,28 @@ int main()
     std::string resultFile{"results/bouncing_ball.csv"};
     const std::string fmuPath{std::string(DATA_FOLDER) + "/fmus/3.0/ref/BouncingBall.fmu"};
 
-    simulation_structure ss;
-    ss.add_model("bouncing_ball", "proxyfmu:://localhost?file=" + fmuPath);
+    try {
+        simulation_structure ss;
+        ss.add_model("bouncing_ball", "proxyfmu:://localhost?file=" + fmuPath);
 
-    const auto sim = ss.load(std::make_unique<fixed_step_algorithm>(1.0 / 60));
-    sim->add_listener("csv_writer", std::make_unique<csv_writer>(resultFile));
+        const auto sim = ss.load(std::make_unique<fixed_step_algorithm>(1.0 / 60));
+        sim->add_listener("csv_writer", std::make_unique<csv_writer>(resultFile));
 
-    sim->init();
-    sim->step_for(10);
-    sim->terminate();
+        sim->init();
+        sim->step_for(10);
+        sim->terminate();
 
-    TChartConfig config;
-    config.addChart(TTimeSeriesChart{
-        "bouncing_ball",
-        "Time[s]",
-        {{{"bouncing_ball", {{"h", }}}}}});
+        TChartConfig config;
+        config.addChart(TTimeSeriesChart{
+            "bouncing_ball",
+            "Time[s]",
+            {{{"bouncing_ball", {{
+                                    "h",
+                                }}}}}});
 
-    plot_csv(resultFile, config);
+        plot_csv(resultFile, config);
+
+    } catch (const std::runtime_error& e) {
+        std::cerr << "Exception raised: " << e.what() << std::endl;
+    }
 }
