+++ src/proxyfmu/process_helper.hpp
@@ -31,7 +31,7 @@ inline std::string getLoc()
 {
 #ifdef __linux__
     char result[PATH_MAX];
-    ssize_t count = readlink("/proc/self/exe", result, sizeof(result)-1);
+    ssize_t count = readlink("/proc/self/exe", result, sizeof(result) - 1);
     if (count != -1) {
         result[count] = '\0'; // Ensure null-termination
         return std::string(result);
@@ -64,7 +64,7 @@ inline void start_process(
 
     if (!exists(executable)) {
         const std::string loc = getLoc();
-        const auto alt_executable = std::filesystem::path(loc).parent_path().string() / executable;
+        const auto alt_executable = std::filesystem::path(loc).parent_path() / executable;
         if (exists(alt_executable)) {
             executable = alt_executable;
         }
@@ -72,19 +72,19 @@ inline void start_process(
 
     std::string execStr = executable.string();
 #ifdef __linux__
-    if (!executable.is_absolute()) {
+    if (is_regular_file(executable) && !executable.is_absolute()) {
         execStr.insert(0, "./");
     }
 #endif
 
-    log::debug("[proxyfmu] Checking if executable is available..");
+    log::debug("[proxyfmu] Checking if executable ({}) is available..", execStr);
     std::ostringstream ss;
     ss << std::quoted(execStr) << " -v";
     const int statusCode = std::system(ss.str().c_str());
     if (statusCode != 0) {
         log::err("[proxyfmu] Unable to invoke proxyfmu!");
 
-        bind.set_value("-");
+        bind.set_value("");
         return;
     }
 
@@ -100,7 +100,7 @@ inline void start_process(
         nullptr};
 
     subprocess_s process{};
-    int result = subprocess_create(cmd.data(), subprocess_option_inherit_environment | subprocess_option_no_window, &process);
+    int result = subprocess_create(cmd.data(), subprocess_option_inherit_environment | subprocess_option_search_user_path | subprocess_option_no_window, &process);
 
     bool bound = false;
     if (result == 0) {
@@ -137,6 +137,9 @@ inline void start_process(
             return;
         }
         log::err("[proxyfmu] External process for instance '{}' returned with status {}. Unable to bind..", instanceName, std::to_string(status));
+    } else {
+
+        log::err("[proxyfmu] Fatal: Subrocess create returned non-zero status: {}", result);
     }
 
     bind.set_value("");
