### 2025-05-14 c0b3320
remove standalone plot script
