### 2025-05-19 c17955e
refactor C/Python string handling
