### 2025-05-19 #8182 @espottesmith
Hey @ElektrikAkar, have you been able to take a look at this paper/code?
