### 2025-05-19 ec13d6c
allow wildcards for CSV config (#57)
