### 2025-05-19 4a00e63
allow setting csv identifiers through C/Python API
