+++ include/ecos/ecos.h
@@ -96,6 +96,7 @@ void ecos_simulation_add_listener(ecos_simulation_t* sim, const char* name, ecos
 void ecos_simulation_remove_listener(ecos_simulation_t* sim, const char* name);
 ecos_simulation_listener_t* ecos_csv_writer_create(const char* resultFile, const char* csvConfig = nullptr);
 bool ecos_csv_writer_set_decimation_factor(ecos_simulation_listener_t* writer, int decimationFactor);
+bool ecos_csv_writer_register_variable(ecos_simulation_listener_t* writer, const char* identifier);
 
 void ecos_plot_csv(const char* csvFile, const char* chartConfig);
 // -------------
