### 2025-05-20 83b46e5
allow global eps for scenario config
