### 2025-05-20 9f4dc0d
make reset pure virtual
