### 2025-05-20 6bd5e60
account for larger step_size with custom decimationFactor
