### 2025-05-20 b907519
expose stepSizeHint in C/Python
