### 2025-05-20 844114f
quarter_truck is windows only [skip ci]
throw exception if os is not windows
