+++ examples/quarter_truck/quarter_truck.py
@@ -3,6 +3,7 @@ from ecospy.plotter import *
 
 from pathlib import Path
 
+
 class MyListener(SimulationListener):
 
     def __init__(self, sim: EcosSimulation):
@@ -24,7 +25,6 @@ def main():
     result_file = f"results/python/quarter_truck.csv"
 
     with (EcosSimulation(ssp_path=f"{ssp_dir}", step_size=1.0 / 100)) as sim:
-
         sim.add_csv_writer(result_file, f"{ssp_dir}/CsvConfig.xml")
         sim.add_listener("custom_listener", MyListener(sim))
 
@@ -51,4 +51,9 @@ def main():
 
 
 if __name__ == "__main__":
+    import sys
+
+    if sys.platform != "win32":
+        raise EnvironmentError("This particular example (quarter_truck) can only be run on Windows.")
+
     main()
