### 2025-05-20 a2bbff4
Fix plot title [skip ci]
