### 2025-05-26 Ecos-platform/ecos#61 @ElektrikAkar
API documentation
Hi, I am reviewing the code from https://github.com/openjournals/joss-reviews/issues/8182

I couldn't see an API documentation, is there a Doxygen-like automated documentation? And/or does the readme.md include all necessary methods?

Many thanks.
