### 2025-05-26 Ecos-platform/ecos#62 @ElektrikAkar
Community guidelines
Hi, I am reviewing the code from https://github.com/openjournals/joss-reviews/issues/8182

As far as I see there is only one contributor. However, is there a guideline for people who wish to contribute to this software? (i.e., `CODE_OF_CONDUCT.md`)
