### 2025-05-26 Ecos-platform/ecos#63 @ElektrikAkar
Statement of need
Hi, I am reviewing the code from https://github.com/openjournals/joss-reviews/issues/8182

Could authors include a short target audience and why ECOS is needed in the documentation/readme.md? It doesn't have to be very long, maybe a copy-paste of the intended audience in the `setup.cfg` and 1-2 sentences from the JOSS paper about ECOS's importance.

Many thanks.
