### 2025-05-26 Ecos-platform/ecos#64 @ElektrikAkar
Comments in the code
Hi, I am reviewing the code from https://github.com/openjournals/joss-reviews/issues/8182

I couldn't see an comments in the code, could authors consider at least adding some comments to the important bits/long functions to make it easier to understand?

Many thanks.
