### 2025-05-26 cbcb9fc
Create CONTRIBUTING.md
