+++ include/ecos/algorithm/algorithm.hpp
@@ -7,7 +7,9 @@
 namespace ecos
 {
 
-/// Abstract base class for algorithms used for co-simulation orchestration.
+/**
+ * /brief Abstract base class for algorithms used for co-simulation orchestration.
+ */
 struct algorithm
 {
 
