+++ ecospy/EcosSimulationRunner.py
@@ -3,8 +3,10 @@ from .EcosSimulation import EcosSimulation
 
 from ctypes import c_void_p, c_double
 
-# This class is used to run a simulation created with EcosSimulation.
 class EcosSimulationRunner:
+    """
+    This class is used to run a simulation.
+    """
 
     def __init__(self, sim: EcosSimulation):
 
