### 2025-05-28 9b4b719
replace Unix domain sockets with Shared memory
