### 2025-05-28 cd4028a
update simple_socket [skip ci]
