### 2025-05-28 Ecos-platform/ecos#61 @elac-safran
Hi @markaren.

**Functionality documentation** is one of the items to be checked for the JOSS review.

I agree with @ElektrikAkar ; this part is currently missing. I will check with the editor what we can do about that.
