### 2025-05-28 Ecos-platform/ecos#69 @elac-safran
FMI v3.0 example
Hi, I am reviewing the code from https://github.com/openjournals/joss-reviews/issues/8182

Congratulations for this nice project.

The JOSS paper mentions that all three versions of the FMI standard are supported by `ecos`.
However, I could not find any example of models with discrete events (introduced in v3.0). Plus, the main algorithm does not seem to implement event detection, time roll back and so on. Have you conducted tests on hybrid continuous/discrete models?
