### 2025-05-28 Ecos-platform/ecos#69 @markaren
_Basic_ support for all three versions. This could be better highlighted. To this day, we have not needed any of the more optional features of the standard. Rollback can be more fleshed out. Discrete events are not handled.

For advances on the orchestration algorithm users need to provide a use-case together with sample systems for verification/test.
