### 2025-05-28 b7bd29b
improve feature descriptions [skip ci]
#61 #69
