### 2025-05-28 a576d7b
Adding documentation (#67)
