### 2025-05-28 Ecos-platform/ecos#69 @elac-safran
Thank you for updating the paper. It seems to me the correct wording would be *partial* support, rather than "basic" support.
Full support can be a future milestone, which is totally fine.
