### 2025-05-30 53ca15e
allow applying parameterset with cli
