### 2025-05-30 5980f23
Allow applying parameterset with cli (#70)
