### 2025-06-03 #8182 @markaren
PRs https://github.com/Ecos-platform/ecos/pull/67,  https://github.com/Ecos-platform/ecos/pull/66 and https://github.com/Ecos-platform/ecos/pull/65  and commit https://github.com/Ecos-platform/ecos/commit/b7bd29b9782bf695fd3673137ec2d913ceef582e  was added in response the the reviewers requests. I hope these changes are in line with the reviewers expectations.
