### 2025-06-05 c9d11cf
Introduce interface types
