### 2025-06-05 0c3a3a8
Add remoting to table
