### 2025-06-06 #8182 @crvernon
:wave: @markaren - I just need you to address the following before I move to accept for publication:

- [ ] LINE 20:  Missing space in between "FMI 3.0" and "(Junghanns..."
- [ ] LINE 27:  "pypy" should be "PyPI"
- [ ] LINE 39: "ip-address" should be "IP address"
- [ ] LINE 44:  Missing space in between "FMPy" and "(CATIA..." as well as "Vico" and "(Hatledal..."
- [ ] LINE 45:  Extra space before the "O" in "( OSP)" and missing one between "OMSimulator" and "(Ochel..."
- [ ] LINE 69:  The "p" in "python" should be capitalized.  You can do this in your bib file by placing characters you wish to maintain formatting of in curly brackets.

No need to conduct a new release for these changes.  Let me know when they are done and I'll accept.  Thanks!
