### 2025-06-06 6006cb4
Fix paper issues based on feedback
