+++ doc/paper/paper.bib
@@ -19,7 +19,7 @@
 }
 
 @misc{FMPy,
-  title        = {FMPy - A Python library to simulate Functional Mock-up Units (FMUs)},
+  title        = {FMPy - A {P}ython library to simulate Functional Mock-up Units (FMUs)},
   author       = {{CATIA Systems}},
   year         = {2017},
   publisher    = {GitHub},
